Move viewer options button to top toolbar
