+++ cubeviz/layout.ui
@@ -495,16 +495,6 @@
            </property>
           </widget>
          </item>
-         <item row="1" column="0">
-          <widget class="QToolButton" name="button_toggle_sidebar">
-           <property name="toolTip">
-            <string>Toggle advanced option panel for Glue data viewers</string>
-           </property>
-           <property name="text">
-            <string>Glue Viewer Options</string>
-           </property>
-          </widget>
-         </item>
         </layout>
        </widget>
       </item>
